Creates a chi.squared test table for multiple dependent variables

Creates a chi.squared test table for multiple dependent variables

chi_test_table(
  dv,
  iv,
  data,
  conditions = levels(factor(iv))[1:2],
  ref_levels = 2,
  labels = NULL,
  nice_p = TRUE,
  digits = 2,
  order = "12",
  label_attr = TRUE,
  type = "html",
  ...
)

Arguments

dv

A data frame with the dependent variables or a character vector when argument data is set.

iv

A data frame or vector with the independent variable or a character if argument data is set.

data

A data frame.

conditions

A character vector of length two with the names of the two conditions. Defaults to the first two levels of the independent variable 'iv' if applicable.

ref_levels

Vector with the reference level for each dependent variable. Values are repeated if length of levels vector is smaller than number dependent variables.

labels

A character vector of length two with labels for the dependent variables.

nice_p

If TRUE, p values are printed in a nice format.

digits

Number of digits for rounding mean and sd values

order

Either "12" or "21" depicting whether group two is compared to group one or vice versa.

type

Either "df" for data frame or "html" for html table.

...

Further arguments passed to [nice_table()]

Value

A tibble or a kableExtra object

Examples

dat <- data.frame(
  gender = factor(rbinom(100, 1, 0.3), labels = c("male", "female")),
  glasses = factor(rbinom(100, 1, 0.3), labels = c("no", "yes")),
  sen = factor(rbinom(100, 1, 0.5), labels = c("no_sen", "sen"))
 )
chi_test_table(
  c("gender", "glasses"), "sen", data = dat, 
  ref_levels = c("male", "no")
)
#> <table class=" lightable-classic-2" style='font-family: "Arial Narrow", "Source Sans Pro", sans-serif; margin-left: auto; margin-right: auto;'>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:left;"> Level </th>
#>    <th style="text-align:left;"> Proportion no_sen </th>
#>    <th style="text-align:left;"> Proportion sen </th>
#>    <th style="text-align:left;"> Difference </th>
#>    <th style="text-align:left;"> X squared </th>
#>    <th style="text-align:right;"> df </th>
#>    <th style="text-align:left;"> p </th>
#>    <th style="text-align:right;"> n no_sen </th>
#>    <th style="text-align:right;"> n sen </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> gender </td>
#>    <td style="text-align:left;"> male </td>
#>    <td style="text-align:left;"> 0.72 </td>
#>    <td style="text-align:left;"> 0.68 </td>
#>    <td style="text-align:left;"> 0.04 </td>
#>    <td style="text-align:left;"> 0.05 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:left;"> .824 </td>
#>    <td style="text-align:right;"> 40 </td>
#>    <td style="text-align:right;"> 60 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> glasses </td>
#>    <td style="text-align:left;"> no </td>
#>    <td style="text-align:left;"> 0.75 </td>
#>    <td style="text-align:left;"> 0.63 </td>
#>    <td style="text-align:left;"> 0.12 </td>
#>    <td style="text-align:left;"> 1.01 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:left;"> .314 </td>
#>    <td style="text-align:right;"> 40 </td>
#>    <td style="text-align:right;"> 60 </td>
#>   </tr>
#> </tbody>
#> </table>