Fisher's Z Transformation

The Fisher's Z transformation is a common method for stabilizing the variance of the sample correlation coefficient.

The inverse Fisher's Z transformation is used to convert a Fisher's Z value back to a correlation coefficient.

fisher_z(r)

inv_fisher_z(z)

Arguments

r

The correlation coefficient to be transformed.

z

The Fisher's Z transformed value.

Value

The Fisher's Z transformed value.

The original correlation coefficient.

Details

The Fisher's Z transformation is defined as: 0.5 * log((1 + r) / (1 - r))

The inverse Fisher's Z transformation is defined as: (exp(2 * z) - 1) / (exp(2 * z) + 1)

References

Fisher, R. A. (1915). Frequency distribution of the values of the correlation coefficient in samples from an indefinitely large population. Biometrika, 10(4), 507-521.

See also

Author

Your Name

Examples

# Example usage:
r <- 0.8
fisher_z(r)
#> [1] 1.098612

# Example usage:
z <- 0.972
inv_fisher_z(z)
#> [1] 0.7495819