Table with alpha values

Returns a data.frame with item analyses for the provided scales.

nice_alpha_table(
  data,
  scales,
  labels = NULL,
  round = 2,
  ci = TRUE,
  conf_level = 0.95,
  check_key = TRUE,
  keys = NULL,
  keys_from_weights = TRUE,
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL,
  fa = TRUE
)

alpha_table(
  data,
  scales,
  labels = NULL,
  round = 2,
  ci = TRUE,
  conf_level = 0.95,
  check_key = TRUE,
  keys = NULL,
  keys_from_weights = TRUE,
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL,
  fa = TRUE
)

Arguments

data

A data Frame

scales

A list containing vectors with variable names. Each list element defines one scale. Named list elements are used as labels.

labels

Label names for scales (defaults to named list elements in 'scales').

round

Rounds values to given decimal position.

ci

If TRUE confidence intervals are calculated.

conf_level

Confidence level (e.g. 0.95 for 95 percent).

check_key

Check_key for the psych::alpha function.

keys

Optional key argument for the psych::alpha function.

keys_from_weights

If TRUE, tries to define keys from scaledics "weights" parameter.

RMSEA

If TRUE RMSEA is calculated.

difficulty

If TRUE, the difficulty of the item is calculated.

values

Sets maximum and minimum valid values necessary to calculate item difficulty

fa

If TRUE, a one factor exploratory factor analyses is calculated and loadings are reported.

Value

A data frame with concise scale indices.

Examples

## Example needs packages scaledic and purrr installed and active
nice_alpha_table(
  data = wmisc:::ex_itrf, 
  scales = wmisc:::ex_itrf_scales, 
  labels = c("Inernalizing", "Externalizing"),
  difficulty = TRUE, 
  values = list(c(0, 3)), 
  RMSEA = TRUE
)
#> Error in names(keys) <- colnames(x): 'names' attribute [19] must be the same length as the vector [0]
#> 
#> Cautions:
#> 1: Removed 1 rows because all items were missing.