Nice correlation matrix

From a psych corr object it gives a correlation matrix formatted for overview

nice_corrmatrix(
  cr,
  upper = FALSE,
  lower = TRUE,
  digits = 2,
  values = TRUE,
  stars = TRUE,
  numbered_columns = TRUE,
  descriptives = TRUE,
  labels = NULL,
  ci = FALSE,
  nsig_p = 0.1,
  char_nsig,
  char_autocor = "-",
  char_p10 = "✝",
  char_NA = "",
  caption = "Correlation matrix.",
  drop_zero = TRUE,
  type = "df",
  ...
)

Arguments

cr

A data frame or a correlation matrix returned by the psych package

upper

TRUE if upper triangle should be included.

lower

TRUE if lower triangle should be included.

digits

Round to given digit position.

values

TRUE if values should be included.

stars

TRUE if stars should be included.

labels

Character string. If "auto" labels are taken from a label attribute.

nsig_p

p level below which correlations are considered not significant.

char_nsig

Character indexing non-significant values.

char_autocor

Character for diagonal (e.g. "-" or "1.00").

char_p10

Character indexing .10 significance level.

char_NA

Character for NA values.

caption

Caption for a html table.

type

Character string. "df" for data-frame. "html" for html table (needs knitr and kableExtra packages).

...

Further arguments passed to the [psych::corr.test()] function.

Value

A data-frame or a html table object

Examples

nice_corrmatrix(mtcars, type = "html", ci = TRUE)
#> <table class=" lightable-classic" style='font-family: "Arial Narrow", "Source Sans Pro", sans-serif; margin-left: auto; margin-right: auto;border-bottom: 0;'>
#> <caption>Correlation matrix.</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:center;"> n </th>
#>    <th style="text-align:center;"> M </th>
#>    <th style="text-align:center;"> SD </th>
#>    <th style="text-align:center;"> 1 </th>
#>    <th style="text-align:center;"> 2 </th>
#>    <th style="text-align:center;"> 3 </th>
#>    <th style="text-align:center;"> 4 </th>
#>    <th style="text-align:center;"> 5 </th>
#>    <th style="text-align:center;"> 6 </th>
#>    <th style="text-align:center;"> 7 </th>
#>    <th style="text-align:center;"> 8 </th>
#>    <th style="text-align:center;"> 9 </th>
#>    <th style="text-align:center;"> 10 </th>
#>    <th style="text-align:center;"> 11 </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> 1. mpg </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 20.09 </td>
#>    <td style="text-align:center;"> 6.03 </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 2. cyl </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 6.19 </td>
#>    <td style="text-align:center;"> 1.79 </td>
#>    <td style="text-align:center;"> -.85***<br>[-.93, -.72] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 3. disp </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 230.72 </td>
#>    <td style="text-align:center;"> 123.94 </td>
#>    <td style="text-align:center;"> -.85***<br>[-.92, -.71] </td>
#>    <td style="text-align:center;"> .90***<br>[.81, .95] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 4. hp </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 146.69 </td>
#>    <td style="text-align:center;"> 68.56 </td>
#>    <td style="text-align:center;"> -.78***<br>[-.89, -.59] </td>
#>    <td style="text-align:center;"> .83***<br>[.68, .92] </td>
#>    <td style="text-align:center;"> .79***<br>[.61, .89] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 5. drat </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 3.60 </td>
#>    <td style="text-align:center;"> 0.53 </td>
#>    <td style="text-align:center;"> .68***<br>[.44, .83] </td>
#>    <td style="text-align:center;"> -.70***<br>[-.84, -.46] </td>
#>    <td style="text-align:center;"> -.71***<br>[-.85, -.48] </td>
#>    <td style="text-align:center;"> -.45** <br>[-.69, -.12] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 6. wt </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 3.22 </td>
#>    <td style="text-align:center;"> 0.98 </td>
#>    <td style="text-align:center;"> -.87***<br>[-.93, -.74] </td>
#>    <td style="text-align:center;"> .78***<br>[.60, .89] </td>
#>    <td style="text-align:center;"> .89***<br>[.78, .94] </td>
#>    <td style="text-align:center;"> .66***<br>[.40, .82] </td>
#>    <td style="text-align:center;"> -.71***<br>[-.85, -.48] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 7. qsec </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 17.85 </td>
#>    <td style="text-align:center;"> 1.79 </td>
#>    <td style="text-align:center;"> .42*  <br>[.08, .67] </td>
#>    <td style="text-align:center;"> -.59***<br>[-.78, -.31] </td>
#>    <td style="text-align:center;"> -.43*  <br>[-.68, -.10] </td>
#>    <td style="text-align:center;"> -.71***<br>[-.85, -.48] </td>
#>    <td style="text-align:center;"> .09   <br>[-.27, .43] </td>
#>    <td style="text-align:center;"> -.17   <br>[-.49, .19] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 8. vs </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 0.44 </td>
#>    <td style="text-align:center;"> 0.50 </td>
#>    <td style="text-align:center;"> .66***<br>[.41, .82] </td>
#>    <td style="text-align:center;"> -.81***<br>[-.90, -.64] </td>
#>    <td style="text-align:center;"> -.71***<br>[-.85, -.48] </td>
#>    <td style="text-align:center;"> -.72***<br>[-.86, -.50] </td>
#>    <td style="text-align:center;"> .44*  <br>[.11, .68] </td>
#>    <td style="text-align:center;"> -.55***<br>[-.76, -.26] </td>
#>    <td style="text-align:center;"> .74***<br>[.53, .87] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 9. am </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 0.41 </td>
#>    <td style="text-align:center;"> 0.50 </td>
#>    <td style="text-align:center;"> .60***<br>[.32, .78] </td>
#>    <td style="text-align:center;"> -.52** <br>[-.74, -.21] </td>
#>    <td style="text-align:center;"> -.59***<br>[-.78, -.31] </td>
#>    <td style="text-align:center;"> -.24   <br>[-.55, .12] </td>
#>    <td style="text-align:center;"> .71***<br>[.48, .85] </td>
#>    <td style="text-align:center;"> -.69***<br>[-.84, -.45] </td>
#>    <td style="text-align:center;"> -.23   <br>[-.54, .13] </td>
#>    <td style="text-align:center;"> .17   <br>[-.19, .49] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 10. gear </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 3.69 </td>
#>    <td style="text-align:center;"> 0.74 </td>
#>    <td style="text-align:center;"> .48** <br>[.16, .71] </td>
#>    <td style="text-align:center;"> -.49** <br>[-.72, -.17] </td>
#>    <td style="text-align:center;"> -.56***<br>[-.76, -.26] </td>
#>    <td style="text-align:center;"> -.13   <br>[-.45, .23] </td>
#>    <td style="text-align:center;"> .70***<br>[.46, .84] </td>
#>    <td style="text-align:center;"> -.58***<br>[-.77, -.29] </td>
#>    <td style="text-align:center;"> -.21   <br>[-.52, .15] </td>
#>    <td style="text-align:center;"> .21   <br>[-.15, .52] </td>
#>    <td style="text-align:center;"> .79***<br>[.62, .89] </td>
#>    <td style="text-align:center;"> - </td>
#>    <td style="text-align:center;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> 11. carb </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 2.81 </td>
#>    <td style="text-align:center;"> 1.62 </td>
#>    <td style="text-align:center;"> -.55** <br>[-.75, -.25] </td>
#>    <td style="text-align:center;"> .53** <br>[.22, .74] </td>
#>    <td style="text-align:center;"> .39*  <br>[.05, .65] </td>
#>    <td style="text-align:center;"> .75***<br>[.54, .87] </td>
#>    <td style="text-align:center;"> -.09   <br>[-.43, .27] </td>
#>    <td style="text-align:center;"> .43*  <br>[.09, .68] </td>
#>    <td style="text-align:center;"> -.66***<br>[-.82, -.40] </td>
#>    <td style="text-align:center;"> -.57***<br>[-.77, -.28] </td>
#>    <td style="text-align:center;"> .06   <br>[-.30, .40] </td>
#>    <td style="text-align:center;"> .27   <br>[-.08, .57] </td>
#>    <td style="text-align:center;"> - </td>
#>   </tr>
#> </tbody>
#> <tfoot>
#> <tr><td style="padding: 0; " colspan="100%"><span style="font-style: italic;">Note: </span></td></tr>
#> <tr><td style="padding: 0; " colspan="100%">
#> <sup></sup> ✝p &lt; .10; *p &lt; .05; **p &lt; .01; ***p &lt; .001</td></tr>
#> </tfoot>
#> </table>