Table with descriptive statistics

Table with descriptive statistics

nice_descriptives(
  data,
  round = NULL,
  auto_labels = FALSE,
  title = "Descriptive statistics",
  footnote = NULL,
  file = NULL
)

Arguments

data

A data frame

round

Digits for round function

auto_labels

If TRUE, variable names are taken from a label attribute.

Value

A data frame with descriptive statistics

Examples

nice_descriptives(mtcars)
#> <div id="elhkpkbjkh" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#elhkpkbjkh table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #elhkpkbjkh thead, #elhkpkbjkh tbody, #elhkpkbjkh tfoot, #elhkpkbjkh tr, #elhkpkbjkh td, #elhkpkbjkh th {
#>   border-style: none;
#> }
#> 
#> #elhkpkbjkh p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #elhkpkbjkh .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #elhkpkbjkh .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #elhkpkbjkh .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #elhkpkbjkh .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #elhkpkbjkh .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #elhkpkbjkh .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #elhkpkbjkh .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #elhkpkbjkh .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #elhkpkbjkh .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #elhkpkbjkh .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #elhkpkbjkh .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #elhkpkbjkh .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #elhkpkbjkh .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #elhkpkbjkh .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #elhkpkbjkh .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #elhkpkbjkh .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #elhkpkbjkh .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #elhkpkbjkh .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #elhkpkbjkh .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #elhkpkbjkh .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #elhkpkbjkh .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #elhkpkbjkh .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #elhkpkbjkh .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #elhkpkbjkh .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #elhkpkbjkh .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #elhkpkbjkh .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #elhkpkbjkh .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #elhkpkbjkh .gt_left {
#>   text-align: left;
#> }
#> 
#> #elhkpkbjkh .gt_center {
#>   text-align: center;
#> }
#> 
#> #elhkpkbjkh .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #elhkpkbjkh .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #elhkpkbjkh .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #elhkpkbjkh .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #elhkpkbjkh .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #elhkpkbjkh .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #elhkpkbjkh .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #elhkpkbjkh .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #elhkpkbjkh .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #elhkpkbjkh .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #elhkpkbjkh .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #elhkpkbjkh .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="10" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Descriptive statistics</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="Variable">Variable</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Valid">Valid</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Missing">Missing</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Mean">Mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="SD">SD</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Min">Min</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Max">Max</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Range">Range</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Median">Median</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="MAD">MAD</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="Variable" class="gt_row gt_left">mpg</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">20.090625</td>
#> <td headers="SD" class="gt_row gt_center">6.0269481</td>
#> <td headers="Min" class="gt_row gt_center">10.400</td>
#> <td headers="Max" class="gt_row gt_center">33.900</td>
#> <td headers="Range" class="gt_row gt_center">23.500</td>
#> <td headers="Median" class="gt_row gt_center">19.200</td>
#> <td headers="MAD" class="gt_row gt_center">5.4114900</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">cyl</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">6.187500</td>
#> <td headers="SD" class="gt_row gt_center">1.7859216</td>
#> <td headers="Min" class="gt_row gt_center">4.000</td>
#> <td headers="Max" class="gt_row gt_center">8.000</td>
#> <td headers="Range" class="gt_row gt_center">4.000</td>
#> <td headers="Median" class="gt_row gt_center">6.000</td>
#> <td headers="MAD" class="gt_row gt_center">2.9652000</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">disp</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">230.721875</td>
#> <td headers="SD" class="gt_row gt_center">123.9386938</td>
#> <td headers="Min" class="gt_row gt_center">71.100</td>
#> <td headers="Max" class="gt_row gt_center">472.000</td>
#> <td headers="Range" class="gt_row gt_center">400.900</td>
#> <td headers="Median" class="gt_row gt_center">196.300</td>
#> <td headers="MAD" class="gt_row gt_center">140.4763500</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">hp</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">146.687500</td>
#> <td headers="SD" class="gt_row gt_center">68.5628685</td>
#> <td headers="Min" class="gt_row gt_center">52.000</td>
#> <td headers="Max" class="gt_row gt_center">335.000</td>
#> <td headers="Range" class="gt_row gt_center">283.000</td>
#> <td headers="Median" class="gt_row gt_center">123.000</td>
#> <td headers="MAD" class="gt_row gt_center">77.0952000</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">drat</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">3.596563</td>
#> <td headers="SD" class="gt_row gt_center">0.5346787</td>
#> <td headers="Min" class="gt_row gt_center">2.760</td>
#> <td headers="Max" class="gt_row gt_center">4.930</td>
#> <td headers="Range" class="gt_row gt_center">2.170</td>
#> <td headers="Median" class="gt_row gt_center">3.695</td>
#> <td headers="MAD" class="gt_row gt_center">0.7042350</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">wt</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">3.217250</td>
#> <td headers="SD" class="gt_row gt_center">0.9784574</td>
#> <td headers="Min" class="gt_row gt_center">1.513</td>
#> <td headers="Max" class="gt_row gt_center">5.424</td>
#> <td headers="Range" class="gt_row gt_center">3.911</td>
#> <td headers="Median" class="gt_row gt_center">3.325</td>
#> <td headers="MAD" class="gt_row gt_center">0.7672455</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">qsec</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">17.848750</td>
#> <td headers="SD" class="gt_row gt_center">1.7869432</td>
#> <td headers="Min" class="gt_row gt_center">14.500</td>
#> <td headers="Max" class="gt_row gt_center">22.900</td>
#> <td headers="Range" class="gt_row gt_center">8.400</td>
#> <td headers="Median" class="gt_row gt_center">17.710</td>
#> <td headers="MAD" class="gt_row gt_center">1.4158830</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">vs</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">0.437500</td>
#> <td headers="SD" class="gt_row gt_center">0.5040161</td>
#> <td headers="Min" class="gt_row gt_center">0.000</td>
#> <td headers="Max" class="gt_row gt_center">1.000</td>
#> <td headers="Range" class="gt_row gt_center">1.000</td>
#> <td headers="Median" class="gt_row gt_center">0.000</td>
#> <td headers="MAD" class="gt_row gt_center">0.0000000</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">am</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">0.406250</td>
#> <td headers="SD" class="gt_row gt_center">0.4989909</td>
#> <td headers="Min" class="gt_row gt_center">0.000</td>
#> <td headers="Max" class="gt_row gt_center">1.000</td>
#> <td headers="Range" class="gt_row gt_center">1.000</td>
#> <td headers="Median" class="gt_row gt_center">0.000</td>
#> <td headers="MAD" class="gt_row gt_center">0.0000000</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">gear</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">3.687500</td>
#> <td headers="SD" class="gt_row gt_center">0.7378041</td>
#> <td headers="Min" class="gt_row gt_center">3.000</td>
#> <td headers="Max" class="gt_row gt_center">5.000</td>
#> <td headers="Range" class="gt_row gt_center">2.000</td>
#> <td headers="Median" class="gt_row gt_center">4.000</td>
#> <td headers="MAD" class="gt_row gt_center">1.4826000</td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">carb</td>
#> <td headers="Valid" class="gt_row gt_center">32</td>
#> <td headers="Missing" class="gt_row gt_center">0</td>
#> <td headers="Mean" class="gt_row gt_center">2.812500</td>
#> <td headers="SD" class="gt_row gt_center">1.6152000</td>
#> <td headers="Min" class="gt_row gt_center">1.000</td>
#> <td headers="Max" class="gt_row gt_center">8.000</td>
#> <td headers="Range" class="gt_row gt_center">7.000</td>
#> <td headers="Median" class="gt_row gt_center">2.000</td>
#> <td headers="MAD" class="gt_row gt_center">1.4826000</td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="10"> <em>Note.</em> MAD is the median average deviation with a consistency adjustment.</td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>