Table with frequency statistics

Table with frequency statistics

nice_frequencies(
  data,
  grouping = NULL,
  label = NULL,
  label_grouping = NULL,
  show_missing = TRUE,
  show_percent = TRUE,
  auto_labels = TRUE,
  title = NULL,
  footnote = NULL,
  file = NULL,
  ...
)

Arguments

data

Vector.

grouping

A grouping variable as a vector.

label

Set label for the variable name.

show_missing

If TRUE, adds a row for the number of missing values.

show_percent

If TRUE, adds a column for percentages.

auto_labels

If TRUE, variable names are taken from a label attribute.

title

Table title.

footnote

Table footnote.

file

Filename.

...

Further arguments passed to nice_table().

Value

An html table with frequencies

Examples

nice_frequencies(mtcars[[11]])
#> <div id="fyoqaitrxp" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#fyoqaitrxp table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #fyoqaitrxp thead, #fyoqaitrxp tbody, #fyoqaitrxp tfoot, #fyoqaitrxp tr, #fyoqaitrxp td, #fyoqaitrxp th {
#>   border-style: none;
#> }
#> 
#> #fyoqaitrxp p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #fyoqaitrxp .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #fyoqaitrxp .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #fyoqaitrxp .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #fyoqaitrxp .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #fyoqaitrxp .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #fyoqaitrxp .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #fyoqaitrxp .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #fyoqaitrxp .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #fyoqaitrxp .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #fyoqaitrxp .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #fyoqaitrxp .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #fyoqaitrxp .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #fyoqaitrxp .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #fyoqaitrxp .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #fyoqaitrxp .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #fyoqaitrxp .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fyoqaitrxp .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #fyoqaitrxp .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #fyoqaitrxp .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #fyoqaitrxp .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fyoqaitrxp .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #fyoqaitrxp .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fyoqaitrxp .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #fyoqaitrxp .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #fyoqaitrxp .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fyoqaitrxp .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #fyoqaitrxp .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #fyoqaitrxp .gt_left {
#>   text-align: left;
#> }
#> 
#> #fyoqaitrxp .gt_center {
#>   text-align: center;
#> }
#> 
#> #fyoqaitrxp .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #fyoqaitrxp .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #fyoqaitrxp .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #fyoqaitrxp .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #fyoqaitrxp .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #fyoqaitrxp .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #fyoqaitrxp .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #fyoqaitrxp .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #fyoqaitrxp .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #fyoqaitrxp .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #fyoqaitrxp .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #fyoqaitrxp .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="3" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Frequency statistics of ‘mtcars[[11]]’</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="mtcars[[11]]">mtcars[[11]]</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Frequency">Frequency</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Percent">Percent</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">1</td>
#> <td headers="Frequency" class="gt_row gt_center">7</td>
#> <td headers="Percent" class="gt_row gt_center">22</td></tr>
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">2</td>
#> <td headers="Frequency" class="gt_row gt_center">10</td>
#> <td headers="Percent" class="gt_row gt_center">31</td></tr>
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">3</td>
#> <td headers="Frequency" class="gt_row gt_center">3</td>
#> <td headers="Percent" class="gt_row gt_center">9</td></tr>
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">4</td>
#> <td headers="Frequency" class="gt_row gt_center">10</td>
#> <td headers="Percent" class="gt_row gt_center">31</td></tr>
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">6</td>
#> <td headers="Frequency" class="gt_row gt_center">1</td>
#> <td headers="Percent" class="gt_row gt_center">3</td></tr>
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">8</td>
#> <td headers="Frequency" class="gt_row gt_center">1</td>
#> <td headers="Percent" class="gt_row gt_center">3</td></tr>
#>     <tr><td headers="mtcars[[11]]" class="gt_row gt_left">Missing</td>
#> <td headers="Frequency" class="gt_row gt_center">0</td>
#> <td headers="Percent" class="gt_row gt_center">0</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>
nice_frequencies(
  mtcars$cyl, mtcars$am, 
  label = "Cylinders", 
  label_grouping = "Automatic"
)
#> <div id="kzgfujivnx" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#kzgfujivnx table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #kzgfujivnx thead, #kzgfujivnx tbody, #kzgfujivnx tfoot, #kzgfujivnx tr, #kzgfujivnx td, #kzgfujivnx th {
#>   border-style: none;
#> }
#> 
#> #kzgfujivnx p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #kzgfujivnx .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #kzgfujivnx .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #kzgfujivnx .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #kzgfujivnx .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #kzgfujivnx .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #kzgfujivnx .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #kzgfujivnx .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #kzgfujivnx .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #kzgfujivnx .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #kzgfujivnx .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #kzgfujivnx .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #kzgfujivnx .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #kzgfujivnx .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #kzgfujivnx .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #kzgfujivnx .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #kzgfujivnx .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzgfujivnx .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #kzgfujivnx .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #kzgfujivnx .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #kzgfujivnx .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzgfujivnx .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #kzgfujivnx .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzgfujivnx .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #kzgfujivnx .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #kzgfujivnx .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzgfujivnx .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzgfujivnx .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzgfujivnx .gt_left {
#>   text-align: left;
#> }
#> 
#> #kzgfujivnx .gt_center {
#>   text-align: center;
#> }
#> 
#> #kzgfujivnx .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #kzgfujivnx .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #kzgfujivnx .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #kzgfujivnx .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #kzgfujivnx .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #kzgfujivnx .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #kzgfujivnx .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #kzgfujivnx .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #kzgfujivnx .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #kzgfujivnx .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #kzgfujivnx .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #kzgfujivnx .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="5" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Frequency statistics of ‘Cylinders’ by ‘Automatic’</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" scope="col" id="Cylinders">Cylinders</th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="Automatic">
#>         <span class="gt_column_spanner">Automatic</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="Percent">
#>         <span class="gt_column_spanner">Percent</span>
#>       </th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="0">0</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="1">1</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id=" 0 "> 0 </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id=" 1 "> 1 </th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="Cylinders" class="gt_row gt_left">4</td>
#> <td headers="0" class="gt_row gt_center">3</td>
#> <td headers="1" class="gt_row gt_center">8</td>
#> <td headers=" 0 " class="gt_row gt_center">16</td>
#> <td headers=" 1 " class="gt_row gt_center">62</td></tr>
#>     <tr><td headers="Cylinders" class="gt_row gt_left">6</td>
#> <td headers="0" class="gt_row gt_center">4</td>
#> <td headers="1" class="gt_row gt_center">3</td>
#> <td headers=" 0 " class="gt_row gt_center">21</td>
#> <td headers=" 1 " class="gt_row gt_center">23</td></tr>
#>     <tr><td headers="Cylinders" class="gt_row gt_left">8</td>
#> <td headers="0" class="gt_row gt_center">12</td>
#> <td headers="1" class="gt_row gt_center">2</td>
#> <td headers=" 0 " class="gt_row gt_center">63</td>
#> <td headers=" 1 " class="gt_row gt_center">15</td></tr>
#>     <tr><td headers="Cylinders" class="gt_row gt_left">Missing</td>
#> <td headers="0" class="gt_row gt_center">0</td>
#> <td headers="1" class="gt_row gt_center">0</td>
#> <td headers=" 0 " class="gt_row gt_center">0</td>
#> <td headers=" 1 " class="gt_row gt_center">0</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>