Table with detailed item statistics for one scale

Returns a data.frame with detailed item analyses for the provided scale.

nice_item_analysis(
  data,
  scales,
  labels = NULL,
  round = 2,
  ci = FALSE,
  conf_level = 0.95,
  check_key = TRUE,
  keys = NULL,
  keys_from_weights = FALSE,
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL,
  fa = TRUE,
  ...
)

Arguments

data

A data Frame

scales

A list with vectors with variable names that define each scale.

labels

Optional labels for the items in the scale.

round

Rounds values to given decimal position.

ci

If TRUE confidence intervals are calculated.

conf_level

Confidence level (e.g. 0.95 for 95 percent).

check_key

Check_key for the psych::alpha function.

keys

Optional key argument for the psych::alpha function.

keys_from_weights

If TRUE, tries to define keys from the weights dictionary attribute. These are only available when you defined them with the scaledic package.

RMSEA

If TRUE RMSEA is calculated.

difficulty

If TRUE, the difficulty of the item is calculated.

values

Sets maximum and minimum valid values necessary to calculate item difficulty.

fa

If TRUE, a one factor exploratory factor analyses is calculated and loadings are reported.

...

Further arguments passed to the nice_table() function.

Value

A data frame with concise scale indices.

Examples

nice_item_analysis(
  wmisc:::data_emo, 
  scale = wmisc:::data_emo_scales, 
  difficulty = TRUE, 
  values = c(1,5)
)

  

Table
Item analysis
    

Labels
      	M
      	SD
      	rit
      	Loadings
      	Difficulty
    

Grundeinstellung
n = 146; Alpha = .68; Std. alpha = .68; Homogeneity = .26
    

F2
	2.54
	1.15
	0.60
	0.83
	0.39

F3
	1.86
	1.11
	0.29
	0.42
	0.21

F26-
	2.58
	1.08
	0.30
	-0.33
	0.39

F27
	2.49
	1.10
	0.41
	0.57
	0.37

F28
	2.90
	1.10
	0.48
	0.51
	0.47

F29
	3.23
	0.92
	0.37
	0.39
	0.56

Resilienz
n = 146; Alpha = .60; Std. alpha = .60; Homogeneity = .23
    

F6
	2.40
	1.06
	0.30
	0.41
	0.35

F10
	1.98
	1.17
	0.37
	0.49
	0.24

F14
	2.29
	1.06
	0.40
	0.56
	0.32

F25-
	1.72
	1.13
	0.34
	-0.45
	0.18

F32-
	2.49
	1.18
	0.36
	-0.49
	0.37

Intuition
n = 145; Alpha = .80; Std. alpha = .81; Homogeneity = .41
    

F1
	3.00
	0.97
	0.60
	0.70
	0.50

F4
	2.85
	0.96
	0.68
	0.78
	0.46

F5
	3.08
	0.93
	0.63
	0.71
	0.52

F16
	3.20
	0.73
	0.50
	0.58
	0.55

F20-
	2.99
	0.99
	0.41
	-0.45
	0.50

F23
	2.79
	1.05
	0.57
	0.63
	0.45

Selbstwahrnehmung
n = 146; Alpha = .72; Std. alpha = .72; Homogeneity = .39
    

F11
	2.47
	1.09
	0.45
	0.53
	0.37

F17
	2.77
	0.97
	0.63
	0.83
	0.44

F18
	2.78
	0.95
	0.48
	0.59
	0.44

F31-
	2.76
	0.98
	0.47
	-0.57
	0.44

Sensibilität
n = 146; Alpha = .57; Std. alpha = .58; Homogeneity = .21
    

F8
	2.58
	0.91
	0.20
	0.29
	0.39

F12
	2.66
	0.95
	0.40
	0.56
	0.42

F21
	2.68
	0.87
	0.45
	0.65
	0.42

F22-
	3.27
	0.99
	0.38
	-0.50
	0.57

F30-
	2.59
	1.05
	0.25
	-0.34
	0.40

Aufmerksamkeit
n = 146; Alpha = .80; Std. alpha = .80; Homogeneity = .41
    

F7
	2.00
	1.09
	0.55
	0.62
	0.25

F9-
	1.41
	1.01
	0.58
	-0.65
	0.10

F13
	2.23
	1.25
	0.52
	0.59
	0.31

F15
	2.20
	1.20
	0.69
	0.81
	0.30

F19
	1.71
	1.05
	0.52
	0.58
	0.18

F24
	1.65
	1.15
	0.51
	0.57
	0.16