Creates table of loadings of exploratory factor analysis based on the psych::fa function

Creates table of loadings of exploratory factor analysis based on the psych::fa function

nice_loadings(
  x,
  factor_names = NULL,
  sort = TRUE,
  cut = 0.2,
  round = 2,
  title = "Loading matrix",
  footnote = NULL,
  auto_labels = TRUE
)

nice_efa(..., file = NULL)

Arguments

x

Object returned from the psych::fa function.

factor_names

A character vector with names for the resulting factors. If not provided, default names are chosen.

sort

If TRUE, loadings are sorted.

cut

Loadings below cut will be omitted in the resulting data frame.

round

Number of digits to round loadings (based on the base::round function)

Value

A data.frame

Examples

wmisc:::mtcars_labeled |> 
  rename_from_labels() |>
  psych::fa(nfactors = 2) |> 
  nice_efa()
#> Loading required namespace: GPArotation

  

Table
Loading matrix
    

Variables
      	
        
Factors

      
      	Communalities
      	Complexity
    

MR1
      	MR2
    

Loadings
    

1/4 mile time
	-0.92
	-0.35
	0.78
	1.29

Gross horsepower
	0.89
	
	0.88
	1.04

Number of carburetors
	0.85
	0.24
	0.66
	1.16

Engine (0 = V-shaped, 1 = straight)
	-0.8
	
	0.71
	1.05

Number of cylinders
	0.71
	-0.48
	0.94
	1.75

Miles/(US) gallon
	-0.61
	0.54
	0.85
	1.98

Transmission (0 = automatic, 1 = manual)
	
	0.93
	0.82
	1.04

Number of forward gears
	0.22
	0.93
	0.8
	1.11

Rear axle ratio
	
	0.78
	0.71
	1.1

Weight (1000 lbs)
	0.42
	-0.7
	0.82
	1.64

Displacement (cu.in.)
	0.58
	-0.6
	0.9
	2

Variances
    

Eigenvalues
	4.71
	4.15
	

Explained variance
	0.43
	0.38
	

Cumulative explained variance
	0.43
	0.81
	

Proportion explained variance
	0.53
	0.47
	

Cumulative proportion explained variance
	0.53
	1
	

 Note. Extraction method is minres. Rotation method is oblimin. RMSEA is 0.178 CI90% [0.121, 0.245]. Loadings below |0.2| are not displayed.
    

nice_efa(
  wmisc:::data_emo_fa,
  factor_names = c(
    "Emotionserkennung", "Resilienz", 
    "Aufmerksamkeit", "Erklärung für Emotionen"
  )
)

  

Table
Loading matrix
    

Variables
      	
        
Factors

      
      	Communalities
      	Complexity
    

Emotionserkennung
      	Resilienz
      	Aufmerksamkeit
      	Erklärung für Emotionen
    

Loadings
    

F5: Ich erkenne sofort, wenn jemand in meiner Umgebung schlechte Laune hat.
	0.7
	
	
	
	0.49
	1.03

F28: Über ein positives Erlebnis freue ich mich noch Tage später.
	0.66
	
	
	-0.22
	0.49
	1.33

F4: Ich erkenne meinem Gegenüber schnell an, wie es ihm gerade geht.
	0.66
	
	
	0.21
	0.52
	1.21

F1: Ich bemerke leicht subtile Veränderungen im Verhalten anderer, die auf deren Gefühle hindeuten.
	0.65
	
	
	
	0.42
	1.02

F29: Wenn ich ein Geschenk erhalte, kann ich mich auch noch Tage später darüber freuen.
	0.64
	
	
	-0.31
	0.44
	1.49

F16: Ich spüre schnell, wenn die Atmosphäre in einer Gruppe angespannt oder harmonisch ist.
	0.61
	
	
	
	0.39
	1.04

F23: Mir fällt es leicht, Emotionen und Gedanken meines Gegenübers zu deuten und darauf empathisch zu reagieren.
	0.57
	
	
	0.2
	0.44
	1.33

F22: Meine Freunde beschreiben mich als unsensibel.
	-0.53
	
	
	
	0.29
	1.17

F21: In sozialen Kontexten erkenne ich auch unausgesprochene Konflikte und Spannungen.
	0.48
	
	
	0.31
	0.36
	1.87

F12: Ich kann Situationen gut einschätzen und weiß, wie ich angemessen reagieren sollte.
	0.47
	0.26
	
	0.25
	0.44
	2.17

F26: Nach schönen Momenten fühle ich mich schnell wieder neutral oder gleichgültig.
	-0.43
	
	
	0.2
	0.21
	1.56

F20: In meinen zwischenmenschlichen Beziehungen kommt es oft zu Unstimmigkeiten, da ich die Emotionen meines Gegenübers nicht nachvollziehen kann.
	-0.39
	
	
	
	0.23
	1.6

F11: Ich kann oft klar benennen, was ich in einem Moment fühle.
	0.36
	
	
	0.35
	0.34
	2.17

F30: Wenn ich glücklich bin und auf eine Person treffe, die traurig ist, fällt es mir schwer, meine Emotionen in den Hintergrund zu stellen und mich ihrer Stimmung anzupassen.
	-0.35
	
	0.24
	
	0.22
	2.43

F27: Selbst wenn ich kurzfristige Rückschläge erlebe, bleibe ich zuversichtlich.
	
	0.78
	
	
	0.61
	1.02

F2: Ich beschreibe mich als optimistischen Menschen.
	0.21
	0.75
	
	
	0.62
	1.18

F3: Ich bin immer gut gelaunt, auch nach stressigen Situationen.
	
	0.65
	
	
	0.4
	1.05

F32: Wenn mein Morgen stressig war, bin ich den ganzen Tag gereizt.
	
	-0.52
	
	
	0.33
	1.22

F8: Ich kann mich gut an neue Gegebenheiten anpassen.
	
	0.46
	
	
	0.26
	1.17

F10: Ich kann mich von Situationen distanzieren, in denen ich starke Emotionen wie Wut und Traurigkeit erlebt habe.
	
	0.46
	
	0.27
	0.3
	1.79

F14: Ich lasse mich von kleinen Alltagssituationen nicht stark belasten.
	
	0.41
	-0.26
	
	0.3
	1.92

F6: Ich fühle mich in der Lage, mit Stresssituationen effektiv umzugehen.
	
	0.39
	
	
	0.24
	1.45

F25: Nach einem Streit brauche ich erstmal viel Zeit für mich, um meine Gefühle zu verarbeiten.
	
	-0.38
	
	
	
	1.54

F15: Ich schweife oft mit meinen Gedanken ab, wenn ich mich eigentlich Konzentrieren sollte.
	
	
	0.8
	
	0.64
	1.04

F9: Ich kann mich gut auf eine Aufgabe fokussieren.
	
	
	-0.63
	
	0.42
	1.08

F13: Ich lasse mich bei Aufgaben oft von meinem Handy ablenken.
	
	
	0.61
	
	0.36
	1.09

F19: In Gesprächen mit anderen schweifen meine Gedanken schnell ab.
	
	
	0.6
	
	0.4
	1.19

F7: Ich fühle mich oft überwältigt, weil es mir schwer fällt, Ablenkungen zu ignorieren.
	
	
	0.59
	
	0.41
	1.17

F24: Mir fällt es schwer, mich auf ein Gespräch zu konzentrieren und mich von anderen Reizen nicht ablenken zu lassen.
	
	
	0.57
	
	0.38
	1.12

F17: Ich verstehe schnell den Grund meiner Emotionen.
	
	
	
	0.63
	0.4
	1.02

F31: Wenn ich starke Emotionen spüre, kann ich selbst nur schwierig einordnen, woher diese kommen.
	
	
	
	-0.61
	0.44
	1.14

F18: Ich weiß oft im Voraus, welche Situationen bestimmte Emotionen in mir auslösen.
	0.22
	
	
	0.47
	0.31
	1.56

Variances
    

Eigenvalues
	4.57
	3.07
	2.75
	1.89
	

Explained variance
	0.14
	0.1
	0.09
	0.06
	

Cumulative explained variance
	0.14
	0.24
	0.32
	0.38
	

Proportion explained variance
	0.37
	0.25
	0.22
	0.15
	

Cumulative proportion explained variance
	0.37
	0.62
	0.85
	1
	

 Note. Extraction method is minres. Rotation method is oblimin. RMSEA is 0.079 CI90% [0.07, 0.088]. Loadings below |0.2| are not displayed.