Create a nice table from one or more regression models

Create a nice table from one or more regression models

nice_regression_table(
  ...,
  round = 2,
  labels_models = NULL,
  rename_labels = list(),
  rename_cols = list(),
  remove_cols = NULL,
  auto_col_names = TRUE,
  file = NULL,
  or = FALSE
)

Arguments

remove_cols

Either column number or column names to be removed.

Examples

lm(mpg ~ am + disp + hp, data = mtcars) |> 
  nice_regression_table()
#> Error in get_wmisc_attributes(x): could not find function "get_wmisc_attributes"
  
nice_regression_table(
  nlme::lme(mpg~disp, data = mtcars, random = ~1|am),
  nlme::lme(mpg~disp + hp, data = mtcars, random = ~1|am)
)   
#> Random effect variances not available. Returned R2 does not account for random effects.
#> Error in get_wmisc_attributes(x): could not find function "get_wmisc_attributes"
  
nice_regression_table(
  wmisc:::model_lmer_1, wmisc:::model_lmer_2,
  rename_labels = list(
    "EffectTrend" = "Trend", "EffectSlope" = "Slope", "TimePost" = "Post", 
    "ConditionTraining" = "Training", "id_subject" = "Subject"),
  rename_cols = list("Estimate" = "B", "SE" = "se"),
  labels_models = c("Only pretest", "Pre- and posttest") 
)
#> Loading required package: lmerTest
#> Loading required package: lme4
#> Loading required package: Matrix
#> 
#> Attaching package: ‘lmerTest’
#> The following object is masked from ‘package:lme4’:
#> 
#>     lmer
#> The following object is masked from ‘package:stats’:
#> 
#>     step
#> Error in get_wmisc_attributes(x): could not find function "get_wmisc_attributes"