Create a nicely formatted table

This function takes a data frame and formats it into a nicely formatted HTML table using the gt packages.

nice_table(
  x,
  title = NULL,
  footnote = NULL,
  spanner = header,
  row_group = NULL,
  row_group_order = NULL,
  header = NULL,
  pack = NULL,
  rownames = FALSE,
  file = NULL,
  cols_label = NULL,
  use_labels = TRUE,
  decimals = NULL,
  round = NULL,
  label_na = NULL,
  engine = getOption("wmisc.nice.table.engine"),
  extra = NULL,
  gt = NULL,
  kable = NULL
)

Arguments

x

The data frame to be formatted into a table

title

Title string.

footnote

Add footnote

row_group

List with information on grouping rows row_group = list("Start" = 1:2, "That is the second" = 3:5)

row_group_order

List with information on grouping order.

file

(works only when engine = "gt") If set, an additional file with the table is produced.

cols_label

List with renaming information for columns (old_name = new_name).

decimals

Number of decimals that will be printed.

round

Number of digits to which numbers should be rounded.

extra

Additional arguments passed to kableExtra::kable_classic_2()

gt

Additional arguments passed to gt::gt()

kable

Additional arguments passed to knitr::kable()

Value

A nicely formatted HTML table

Examples

df <- data.frame(
  x = 1:5, y = rnorm(5, mean = 10, sd = 12),
  c = letters[5:1], d = sample(letters, 5)
)
nice_table(
  df,
  title = "A nice title",
  footnote = c("Footnote 1", "Footnote 2"),
  spanner = list("One" = 1:2, "Two" = 3:4),
  row_group = list("Start" = 1:2, "That is the second" = 3:5),
  cols_label = list(x = "First", y = "Second", c = "Third", d = "Fourth"),
  decimals = 1
)
#> Error in get_wmisc_attributes(x): could not find function "get_wmisc_attributes"