Creates a t.test table for multiple dependent variables

Creates a t.test table for multiple dependent variables

t_test_table(
  dv,
  iv,
  data,
  method = "cohen",
  conditions = levels(factor(iv))[order],
  labels = NULL,
  concise = TRUE,
  nice_p = TRUE,
  digits = 1,
  var_equal = FALSE,
  label_attr = TRUE,
  manova = TRUE,
  order = 1:2,
  type = "html",
  caption = "t-test table",
  bootstrap_options = c("condensed", "striped", "hover"),
  full_width = TRUE
)

Arguments

dv

A data frame with the dependent variables or a character vector with variable names when data is defined.

iv

A data frame or vector with the independent variable or a character if data is defined.

data

A data frame.

conditions

A character vector of length two with the names of the two conditions. Defaults to the first two levels of the independent variable 'iv' if applicable.

labels

A character vector of length two with labels for the dependent variables.

concise

A more concise table with mean and SD in one column.

nice_p

If TRUE, p values are printed in a nice format.

digits

Number of digits for rounding mean and SD values

var_equal

If FALSE, a t-test for unequal variances is calculated.

order

Either "12" or "21" depicting whether group two is compared to group one or vice versa.

type

Either "df" for data frame or "html" for html table.

caption

Table caption is type = "html"

bootstrap_options

see kable_styling()

full_width

see kable_styling()

Value

A tibble or a kableExtra object

Examples

dv <- data.frame(
  a = c(rnorm(85, 50, 10), rnorm(200, 70, 20)),
  b = c(rnorm(85, 50, 10), rnorm(200, 55, 20)),
  iv = factor(c(rep("A", 85), rep("B", 200)))
)

t_test_table(
  c("a", "b"), "iv", data = dv, 
  labels = c("Motivation", "Achievement")
)
#> <table class=" lightable-classic" style='font-family: "Arial Narrow", "Source Sans Pro", sans-serif; margin-left: auto; margin-right: auto;border-bottom: 0;'>
#> <caption>t-test table</caption>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Scale </th>
#>    <th style="text-align:center;"> M (SD) A </th>
#>    <th style="text-align:center;"> M (SD) B </th>
#>    <th style="text-align:center;"> d </th>
#>    <th style="text-align:center;"> t </th>
#>    <th style="text-align:center;"> df </th>
#>    <th style="text-align:center;"> p </th>
#>    <th style="text-align:center;"> n A </th>
#>    <th style="text-align:center;"> n B </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Motivation </td>
#>    <td style="text-align:center;"> 51.2 ( 9.9) </td>
#>    <td style="text-align:center;"> 68.9 (19.8) </td>
#>    <td style="text-align:center;"> -1.0 </td>
#>    <td style="text-align:center;"> -10.03 </td>
#>    <td style="text-align:center;"> 276.1 </td>
#>    <td style="text-align:center;"> &lt;.001 </td>
#>    <td style="text-align:center;"> 85 </td>
#>    <td style="text-align:center;"> 200 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Achievement </td>
#>    <td style="text-align:center;"> 50.4 (10.8) </td>
#>    <td style="text-align:center;"> 56.5 (21.3) </td>
#>    <td style="text-align:center;"> -0.3 </td>
#>    <td style="text-align:center;"> -3.19 </td>
#>    <td style="text-align:center;"> 274.6 </td>
#>    <td style="text-align:center;"> &lt;.001 </td>
#>    <td style="text-align:center;"> 85 </td>
#>    <td style="text-align:center;"> 200 </td>
#>   </tr>
#> </tbody>
#> <tfoot>
#> <tr><td style="padding: 0; " colspan="100%"><span style="font-style: italic;">Note: </span></td></tr>
#> <tr><td style="padding: 0; " colspan="100%">
#> <sup></sup> Method for estimating standard deviation: cohen; Manova: Pillai = 0.19; F(2, 282) = 33.50; p = 0.000</td></tr>
#> </tfoot>
#> </table>