Nice correlation matrix

From a data.frame object it gives a correlation matrix formatted for overview

nice_corrmatrix(
  cr,
  upper = FALSE,
  lower = TRUE,
  digits = 2,
  show_r = TRUE,
  show_p = FALSE,
  show_stars = TRUE,
  show_ci = FALSE,
  numbered_columns = TRUE,
  show_descriptives = TRUE,
  labels = NULL,
  nsig_p = 0.1,
  char_nsig,
  char_autocor = "－",
  char_p10 = "✞",
  char_NA = "",
  string_ci = "{break_sign}[{ci_lower},{ci_upper}]",
  string_p = "{break_sign}(p {nice_p(p, equal_sign = TRUE)})",
  caption = "Correlation matrix",
  drop_zero = TRUE,
  type = "html",
  file = NULL,
  ...
)

Arguments

cr

A data frame

upper

TRUE if upper triangle should be included.

lower

TRUE if lower triangle should be included.

digits

Round to given digit position.

show_r

TRUE if r-values should be included.

show_p

TRUE if r-values should be included.

show_stars

TRUE if stars should be included.

show_ci

If TRUE, confidence intervals are added.

labels

Character string. If "auto" labels are taken from a label attribute.

nsig_p

p level below which correlations are considered not significant.

char_nsig

Character indexing non-significant values.

char_autocor

Character for diagonal (e.g. "-" or "1.00").

char_p10

Character indexing .10 significance level.

char_NA

Character for NA values.

string_ci

Character string in glue format for confindence internvals.

string_p

Character string in glue format for p-values.

caption

Caption for an html table.

drop_zero

If TRUE, leadning zeros are dropped.

type

Character string. "df" for data-frame. "html" for html table.

file

If TRUE or a filename is provided, a file is exportet (format is defined by file ending eith html or docx).

...

Further arguments passed to the cor.test() function.

show_discriptives

If TRUE, mean and sd columns are added.

Value

A data-frame or a html table object

Examples

nice_corrmatrix(mtcars)
#> <div id="qttfibnusl" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#qttfibnusl table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #qttfibnusl thead, #qttfibnusl tbody, #qttfibnusl tfoot, #qttfibnusl tr, #qttfibnusl td, #qttfibnusl th {
#>   border-style: none;
#> }
#> 
#> #qttfibnusl p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #qttfibnusl .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #qttfibnusl .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #qttfibnusl .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #qttfibnusl .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #qttfibnusl .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #qttfibnusl .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #qttfibnusl .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #qttfibnusl .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #qttfibnusl .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #qttfibnusl .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #qttfibnusl .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #qttfibnusl .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #qttfibnusl .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #qttfibnusl .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #qttfibnusl .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #qttfibnusl .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qttfibnusl .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #qttfibnusl .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #qttfibnusl .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #qttfibnusl .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qttfibnusl .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #qttfibnusl .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qttfibnusl .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #qttfibnusl .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #qttfibnusl .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qttfibnusl .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qttfibnusl .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qttfibnusl .gt_left {
#>   text-align: left;
#> }
#> 
#> #qttfibnusl .gt_center {
#>   text-align: center;
#> }
#> 
#> #qttfibnusl .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #qttfibnusl .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #qttfibnusl .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #qttfibnusl .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #qttfibnusl .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #qttfibnusl .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #qttfibnusl .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #qttfibnusl .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #qttfibnusl .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #qttfibnusl .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #qttfibnusl .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #qttfibnusl .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="15" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Correlation matrix</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="Variable">Variable</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="n">n</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="M">M</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="SD">SD</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="1">1</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="2">2</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="3">3</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="4">4</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="5">5</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="6">6</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="7">7</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="8">8</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="9">9</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="10">10</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="11">11</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="Variable" class="gt_row gt_left">1. mpg  </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center"> 20.09</td>
#> <td headers="SD" class="gt_row gt_center">  6.03</td>
#> <td headers="1" class="gt_row gt_center">－      </td>
#> <td headers="2" class="gt_row gt_center">        </td>
#> <td headers="3" class="gt_row gt_center">        </td>
#> <td headers="4" class="gt_row gt_center">        </td>
#> <td headers="5" class="gt_row gt_center">        </td>
#> <td headers="6" class="gt_row gt_center">        </td>
#> <td headers="7" class="gt_row gt_center">        </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">2. cyl  </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  6.19</td>
#> <td headers="SD" class="gt_row gt_center">  1.79</td>
#> <td headers="1" class="gt_row gt_center">-.85*** </td>
#> <td headers="2" class="gt_row gt_center">－      </td>
#> <td headers="3" class="gt_row gt_center">        </td>
#> <td headers="4" class="gt_row gt_center">        </td>
#> <td headers="5" class="gt_row gt_center">        </td>
#> <td headers="6" class="gt_row gt_center">        </td>
#> <td headers="7" class="gt_row gt_center">        </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">3. disp </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">230.72</td>
#> <td headers="SD" class="gt_row gt_center">123.94</td>
#> <td headers="1" class="gt_row gt_center">-.85*** </td>
#> <td headers="2" class="gt_row gt_center">  .90***</td>
#> <td headers="3" class="gt_row gt_center">－      </td>
#> <td headers="4" class="gt_row gt_center">        </td>
#> <td headers="5" class="gt_row gt_center">        </td>
#> <td headers="6" class="gt_row gt_center">        </td>
#> <td headers="7" class="gt_row gt_center">        </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">4. hp   </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">146.69</td>
#> <td headers="SD" class="gt_row gt_center"> 68.56</td>
#> <td headers="1" class="gt_row gt_center">-.78*** </td>
#> <td headers="2" class="gt_row gt_center">  .83***</td>
#> <td headers="3" class="gt_row gt_center">  .79***</td>
#> <td headers="4" class="gt_row gt_center">－      </td>
#> <td headers="5" class="gt_row gt_center">        </td>
#> <td headers="6" class="gt_row gt_center">        </td>
#> <td headers="7" class="gt_row gt_center">        </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">5. drat </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  3.60</td>
#> <td headers="SD" class="gt_row gt_center">  0.53</td>
#> <td headers="1" class="gt_row gt_center">  .68***</td>
#> <td headers="2" class="gt_row gt_center">-.70*** </td>
#> <td headers="3" class="gt_row gt_center">-.71*** </td>
#> <td headers="4" class="gt_row gt_center">-.45**  </td>
#> <td headers="5" class="gt_row gt_center">－      </td>
#> <td headers="6" class="gt_row gt_center">        </td>
#> <td headers="7" class="gt_row gt_center">        </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">6. wt   </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  3.22</td>
#> <td headers="SD" class="gt_row gt_center">  0.98</td>
#> <td headers="1" class="gt_row gt_center">-.87*** </td>
#> <td headers="2" class="gt_row gt_center">  .78***</td>
#> <td headers="3" class="gt_row gt_center">  .89***</td>
#> <td headers="4" class="gt_row gt_center">  .66***</td>
#> <td headers="5" class="gt_row gt_center">-.71*** </td>
#> <td headers="6" class="gt_row gt_center">－      </td>
#> <td headers="7" class="gt_row gt_center">        </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">7. qsec </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center"> 17.85</td>
#> <td headers="SD" class="gt_row gt_center">  1.79</td>
#> <td headers="1" class="gt_row gt_center">  .42*  </td>
#> <td headers="2" class="gt_row gt_center">-.59*** </td>
#> <td headers="3" class="gt_row gt_center">-.43*   </td>
#> <td headers="4" class="gt_row gt_center">-.71*** </td>
#> <td headers="5" class="gt_row gt_center">  .09   </td>
#> <td headers="6" class="gt_row gt_center">-.17    </td>
#> <td headers="7" class="gt_row gt_center">－      </td>
#> <td headers="8" class="gt_row gt_center">        </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">8. vs   </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  0.44</td>
#> <td headers="SD" class="gt_row gt_center">  0.50</td>
#> <td headers="1" class="gt_row gt_center">  .66***</td>
#> <td headers="2" class="gt_row gt_center">-.81*** </td>
#> <td headers="3" class="gt_row gt_center">-.71*** </td>
#> <td headers="4" class="gt_row gt_center">-.72*** </td>
#> <td headers="5" class="gt_row gt_center">  .44*  </td>
#> <td headers="6" class="gt_row gt_center">-.55*** </td>
#> <td headers="7" class="gt_row gt_center">  .74***</td>
#> <td headers="8" class="gt_row gt_center">－      </td>
#> <td headers="9" class="gt_row gt_center">        </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">9. am   </td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  0.41</td>
#> <td headers="SD" class="gt_row gt_center">  0.50</td>
#> <td headers="1" class="gt_row gt_center">  .60***</td>
#> <td headers="2" class="gt_row gt_center">-.52**  </td>
#> <td headers="3" class="gt_row gt_center">-.59*** </td>
#> <td headers="4" class="gt_row gt_center">-.24    </td>
#> <td headers="5" class="gt_row gt_center">  .71***</td>
#> <td headers="6" class="gt_row gt_center">-.69*** </td>
#> <td headers="7" class="gt_row gt_center">-.23    </td>
#> <td headers="8" class="gt_row gt_center">  .17   </td>
#> <td headers="9" class="gt_row gt_center">－      </td>
#> <td headers="10" class="gt_row gt_center">        </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">10. gear</td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  3.69</td>
#> <td headers="SD" class="gt_row gt_center">  0.74</td>
#> <td headers="1" class="gt_row gt_center">  .48** </td>
#> <td headers="2" class="gt_row gt_center">-.49**  </td>
#> <td headers="3" class="gt_row gt_center">-.56*** </td>
#> <td headers="4" class="gt_row gt_center">-.13    </td>
#> <td headers="5" class="gt_row gt_center">  .70***</td>
#> <td headers="6" class="gt_row gt_center">-.58*** </td>
#> <td headers="7" class="gt_row gt_center">-.21    </td>
#> <td headers="8" class="gt_row gt_center">  .21   </td>
#> <td headers="9" class="gt_row gt_center">  .79***</td>
#> <td headers="10" class="gt_row gt_center">－      </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">11. carb</td>
#> <td headers="n" class="gt_row gt_center">32</td>
#> <td headers="M" class="gt_row gt_center">  2.81</td>
#> <td headers="SD" class="gt_row gt_center">  1.62</td>
#> <td headers="1" class="gt_row gt_center">-.55**  </td>
#> <td headers="2" class="gt_row gt_center">  .53** </td>
#> <td headers="3" class="gt_row gt_center">  .39*  </td>
#> <td headers="4" class="gt_row gt_center">  .75***</td>
#> <td headers="5" class="gt_row gt_center">-.09    </td>
#> <td headers="6" class="gt_row gt_center">  .43*  </td>
#> <td headers="7" class="gt_row gt_center">-.66*** </td>
#> <td headers="8" class="gt_row gt_center">-.57*** </td>
#> <td headers="9" class="gt_row gt_center">  .06   </td>
#> <td headers="10" class="gt_row gt_center">  .27   </td>
#> <td headers="11" class="gt_row gt_center">－</td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="15"> <em>Note.</em> ✞<em>p</em> &lt; .10; *<em>p</em> &lt; .05; **<em>p</em> &lt; .01; ***<em>p</em> &lt; .001.</td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>
nice_corrmatrix(mtcars, 
  show_p = TRUE, 
  show_ci = TRUE, 
  show_stars = FALSE, 
  show_descriptives = FALSE,
  conf.level = 0.99
)
#> <div id="abzoyayyfv" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#abzoyayyfv table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #abzoyayyfv thead, #abzoyayyfv tbody, #abzoyayyfv tfoot, #abzoyayyfv tr, #abzoyayyfv td, #abzoyayyfv th {
#>   border-style: none;
#> }
#> 
#> #abzoyayyfv p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #abzoyayyfv .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #abzoyayyfv .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #abzoyayyfv .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #abzoyayyfv .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #abzoyayyfv .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #abzoyayyfv .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #abzoyayyfv .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #abzoyayyfv .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #abzoyayyfv .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #abzoyayyfv .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #abzoyayyfv .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #abzoyayyfv .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #abzoyayyfv .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #abzoyayyfv .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #abzoyayyfv .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #abzoyayyfv .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #abzoyayyfv .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #abzoyayyfv .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #abzoyayyfv .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #abzoyayyfv .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #abzoyayyfv .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #abzoyayyfv .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #abzoyayyfv .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #abzoyayyfv .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #abzoyayyfv .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #abzoyayyfv .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #abzoyayyfv .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #abzoyayyfv .gt_left {
#>   text-align: left;
#> }
#> 
#> #abzoyayyfv .gt_center {
#>   text-align: center;
#> }
#> 
#> #abzoyayyfv .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #abzoyayyfv .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #abzoyayyfv .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #abzoyayyfv .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #abzoyayyfv .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #abzoyayyfv .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #abzoyayyfv .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #abzoyayyfv .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #abzoyayyfv .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #abzoyayyfv .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #abzoyayyfv .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #abzoyayyfv .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="12" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Correlation matrix</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="Variable">Variable</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="1">1</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="2">2</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="3">3</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="4">4</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="5">5</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="6">6</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="7">7</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="8">8</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="9">9</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="10">10</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="11">11</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="Variable" class="gt_row gt_left">1. mpg  </td>
#> <td headers="1" class="gt_row gt_center">－                                </td>
#> <td headers="2" class="gt_row gt_center">                                  </td>
#> <td headers="3" class="gt_row gt_center">                                  </td>
#> <td headers="4" class="gt_row gt_center">                                  </td>
#> <td headers="5" class="gt_row gt_center">                                  </td>
#> <td headers="6" class="gt_row gt_center">                                  </td>
#> <td headers="7" class="gt_row gt_center">                                  </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">2. cyl  </td>
#> <td headers="1" class="gt_row gt_center">-.85 <br>[-.94,-.66] <br>(p <.001)</td>
#> <td headers="2" class="gt_row gt_center">－                                </td>
#> <td headers="3" class="gt_row gt_center">                                  </td>
#> <td headers="4" class="gt_row gt_center">                                  </td>
#> <td headers="5" class="gt_row gt_center">                                  </td>
#> <td headers="6" class="gt_row gt_center">                                  </td>
#> <td headers="7" class="gt_row gt_center">                                  </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">3. disp </td>
#> <td headers="1" class="gt_row gt_center">-.85 <br>[-.94,-.65] <br>(p <.001)</td>
#> <td headers="2" class="gt_row gt_center">  .90 <br>[.76,.96] <br>(p <.001) </td>
#> <td headers="3" class="gt_row gt_center">－                                </td>
#> <td headers="4" class="gt_row gt_center">                                  </td>
#> <td headers="5" class="gt_row gt_center">                                  </td>
#> <td headers="6" class="gt_row gt_center">                                  </td>
#> <td headers="7" class="gt_row gt_center">                                  </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">4. hp   </td>
#> <td headers="1" class="gt_row gt_center">-.78 <br>[-.91,-.51] <br>(p <.001)</td>
#> <td headers="2" class="gt_row gt_center">  .83 <br>[.62,.93] <br>(p <.001) </td>
#> <td headers="3" class="gt_row gt_center">  .79 <br>[.53,.91] <br>(p <.001) </td>
#> <td headers="4" class="gt_row gt_center">－                                </td>
#> <td headers="5" class="gt_row gt_center">                                  </td>
#> <td headers="6" class="gt_row gt_center">                                  </td>
#> <td headers="7" class="gt_row gt_center">                                  </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">5. drat </td>
#> <td headers="1" class="gt_row gt_center">  .68 <br>[.34,.86] <br>(p <.001) </td>
#> <td headers="2" class="gt_row gt_center">-.70 <br>[-.87,-.37] <br>(p <.001)</td>
#> <td headers="3" class="gt_row gt_center">-.71 <br>[-.88,-.39] <br>(p <.001)</td>
#> <td headers="4" class="gt_row gt_center">-.45 <br>[-.74,-.00] <br>(p <.01) </td>
#> <td headers="5" class="gt_row gt_center">－                                </td>
#> <td headers="6" class="gt_row gt_center">                                  </td>
#> <td headers="7" class="gt_row gt_center">                                  </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">6. wt   </td>
#> <td headers="1" class="gt_row gt_center">-.87 <br>[-.95,-.69] <br>(p <.001)</td>
#> <td headers="2" class="gt_row gt_center">  .78 <br>[.52,.91] <br>(p <.001) </td>
#> <td headers="3" class="gt_row gt_center">  .89 <br>[.73,.96] <br>(p <.001) </td>
#> <td headers="4" class="gt_row gt_center">  .66 <br>[.30,.85] <br>(p <.001) </td>
#> <td headers="5" class="gt_row gt_center">-.71 <br>[-.88,-.39] <br>(p <.001)</td>
#> <td headers="6" class="gt_row gt_center">－                                </td>
#> <td headers="7" class="gt_row gt_center">                                  </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">7. qsec </td>
#> <td headers="1" class="gt_row gt_center">  .42 <br>[-.03,.73] <br>(p <.05) </td>
#> <td headers="2" class="gt_row gt_center">-.59 <br>[-.82,-.20] <br>(p <.001)</td>
#> <td headers="3" class="gt_row gt_center">-.43 <br>[-.74,.01] <br>(p <.05)  </td>
#> <td headers="4" class="gt_row gt_center">-.71 <br>[-.88,-.38] <br>(p <.001)</td>
#> <td headers="5" class="gt_row gt_center">  .09 <br>[-.37,.52] <br>(p =.620)</td>
#> <td headers="6" class="gt_row gt_center">-.17 <br>[-.57,.29] <br>(p =.339) </td>
#> <td headers="7" class="gt_row gt_center">－                                </td>
#> <td headers="8" class="gt_row gt_center">                                  </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">8. vs   </td>
#> <td headers="1" class="gt_row gt_center">  .66 <br>[.31,.86] <br>(p <.001) </td>
#> <td headers="2" class="gt_row gt_center">-.81 <br>[-.92,-.57] <br>(p <.001)</td>
#> <td headers="3" class="gt_row gt_center">-.71 <br>[-.88,-.39] <br>(p <.001)</td>
#> <td headers="4" class="gt_row gt_center">-.72 <br>[-.88,-.41] <br>(p <.001)</td>
#> <td headers="5" class="gt_row gt_center">  .44 <br>[-.01,.74] <br>(p <.05) </td>
#> <td headers="6" class="gt_row gt_center">-.55 <br>[-.80,-.15] <br>(p <.001)</td>
#> <td headers="7" class="gt_row gt_center">  .74 <br>[.45,.89] <br>(p <.001) </td>
#> <td headers="8" class="gt_row gt_center">－                                </td>
#> <td headers="9" class="gt_row gt_center">                                  </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">9. am   </td>
#> <td headers="1" class="gt_row gt_center">  .60 <br>[.21,.82] <br>(p <.001) </td>
#> <td headers="2" class="gt_row gt_center">-.52 <br>[-.78,-.10] <br>(p <.01) </td>
#> <td headers="3" class="gt_row gt_center">-.59 <br>[-.82,-.20] <br>(p <.001)</td>
#> <td headers="4" class="gt_row gt_center">-.24 <br>[-.62,.23] <br>(p =.180) </td>
#> <td headers="5" class="gt_row gt_center">  .71 <br>[.39,.88] <br>(p <.001) </td>
#> <td headers="6" class="gt_row gt_center">-.69 <br>[-.87,-.36] <br>(p <.001)</td>
#> <td headers="7" class="gt_row gt_center">-.23 <br>[-.61,.24] <br>(p =.206) </td>
#> <td headers="8" class="gt_row gt_center">  .17 <br>[-.30,.57] <br>(p =.357)</td>
#> <td headers="9" class="gt_row gt_center">－                                </td>
#> <td headers="10" class="gt_row gt_center">                                  </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">10. gear</td>
#> <td headers="1" class="gt_row gt_center">  .48 <br>[.05,.76] <br>(p <.01)  </td>
#> <td headers="2" class="gt_row gt_center">-.49 <br>[-.77,-.06] <br>(p <.01) </td>
#> <td headers="3" class="gt_row gt_center">-.56 <br>[-.80,-.15] <br>(p <.001)</td>
#> <td headers="4" class="gt_row gt_center">-.13 <br>[-.54,.34] <br>(p =.493) </td>
#> <td headers="5" class="gt_row gt_center">  .70 <br>[.37,.87] <br>(p <.001) </td>
#> <td headers="6" class="gt_row gt_center">-.58 <br>[-.82,-.19] <br>(p <.001)</td>
#> <td headers="7" class="gt_row gt_center">-.21 <br>[-.60,.26] <br>(p =.243) </td>
#> <td headers="8" class="gt_row gt_center">  .21 <br>[-.26,.60] <br>(p =.258)</td>
#> <td headers="9" class="gt_row gt_center">  .79 <br>[.54,.92] <br>(p <.001) </td>
#> <td headers="10" class="gt_row gt_center">－                                </td>
#> <td headers="11" class="gt_row gt_center">  </td></tr>
#>     <tr><td headers="Variable" class="gt_row gt_left">11. carb</td>
#> <td headers="1" class="gt_row gt_center">-.55 <br>[-.80,-.14] <br>(p <.01) </td>
#> <td headers="2" class="gt_row gt_center">  .53 <br>[.11,.79] <br>(p <.01)  </td>
#> <td headers="3" class="gt_row gt_center">  .39 <br>[-.06,.71] <br>(p <.05) </td>
#> <td headers="4" class="gt_row gt_center">  .75 <br>[.46,.90] <br>(p <.001) </td>
#> <td headers="5" class="gt_row gt_center">-.09 <br>[-.51,.37] <br>(p =.621) </td>
#> <td headers="6" class="gt_row gt_center">  .43 <br>[-.02,.73] <br>(p <.05) </td>
#> <td headers="7" class="gt_row gt_center">-.66 <br>[-.85,-.30] <br>(p <.001)</td>
#> <td headers="8" class="gt_row gt_center">-.57 <br>[-.81,-.17] <br>(p <.001)</td>
#> <td headers="9" class="gt_row gt_center">  .06 <br>[-.40,.49] <br>(p =.754)</td>
#> <td headers="10" class="gt_row gt_center">  .27 <br>[-.19,.64] <br>(p =.129)</td>
#> <td headers="11" class="gt_row gt_center">－</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>