Table with detailed item statistics for one scale

Returns a data.frame with detailed item analyses for the provided scale.

nice_item_analysis(
  data,
  scale,
  labels = NULL,
  round = 2,
  ci = TRUE,
  conf_level = 0.95,
  check_key = TRUE,
  keys = NULL,
  keys_from_weights = FALSE,
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL,
  fa = TRUE
)

Arguments

data

A data Frame

scale

A vector with variable names that define the scale.

round

Rounds values to given decimal position.

ci

If TRUE confidence intervals are calculated.

conf_level

Confidence level (e.g. 0.95 for 95 percent).

check_key

Check_key for the psych::alpha function.

keys

Optional key argument for the psych::alpha function.

keys_from_weights

If TRUE, tries to define keys from the weights dictionary attribute. These are only available when you defined them with the scaledic package.

RMSEA

If TRUE RMSEA is calculated.

difficulty

If TRUE, the difficulty of the item is calculated.

values

Sets maximum and minimum valid values necessary to calculate item difficulty.

fa

If TRUE, a one factor exploratory factor analyses is calculated and loadings are reported.

Value

A data frame with concise scale indices.

Examples

nice_item_analysis(
  wmisc:::data_emo, 
  scale = wmisc:::data_emo_scales[[1]], 
  difficulty = TRUE, 
  values = c(1,5)
)

  

Table
Item analysis
    

Labels
      	n
      	M
      	SD
      	Discrimination
      	Loadings
      	Difficulty
    

F2
	146
	2.54
	1.15
	0.60
	0.83
	0.39

F3
	146
	1.86
	1.11
	0.29
	0.42
	0.21

F26-
	145
	2.58
	1.08
	0.30
	-0.33
	0.39

F27
	146
	2.49
	1.10
	0.41
	0.57
	0.37

F28
	146
	2.90
	1.10
	0.48
	0.51
	0.47

F29
	145
	3.23
	0.92
	0.37
	0.39
	0.56