Creates table of loadings of exploratory factor analysis based on the psych::fa function

Creates table of loadings of exploratory factor analysis based on the psych::fa function

nice_loadings(
  x,
  factor_names = NULL,
  sort = TRUE,
  cut = 0.2,
  round = 2,
  title = "Loading matrix",
  footnote = NULL,
  auto_labels = TRUE
)

nice_efa(..., file = NULL)

Arguments

x

Object returned from the psych::fa function.

factor_names

A character vector with names for the resulting factors. If not provided, default names are chosen.

sort

If TRUE, loadings are sorted.

cut

Loadings below cut will be omitted in the resulting data frame.

round

Number of digits to round loadings (based on the base::round function)

Value

A data.frame

Examples

mtcars |> psych::fa(nfactors = 2) |> nice_efa()
#> Loading required namespace: GPArotation
#> <div id="lgxvjtiwcr" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#lgxvjtiwcr table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #lgxvjtiwcr thead, #lgxvjtiwcr tbody, #lgxvjtiwcr tfoot, #lgxvjtiwcr tr, #lgxvjtiwcr td, #lgxvjtiwcr th {
#>   border-style: none;
#> }
#> 
#> #lgxvjtiwcr p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #lgxvjtiwcr .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #lgxvjtiwcr .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #lgxvjtiwcr .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #lgxvjtiwcr .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #lgxvjtiwcr .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #lgxvjtiwcr .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #lgxvjtiwcr .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #lgxvjtiwcr .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #lgxvjtiwcr .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #lgxvjtiwcr .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #lgxvjtiwcr .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #lgxvjtiwcr .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #lgxvjtiwcr .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #lgxvjtiwcr .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #lgxvjtiwcr .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #lgxvjtiwcr .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #lgxvjtiwcr .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #lgxvjtiwcr .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #lgxvjtiwcr .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #lgxvjtiwcr .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #lgxvjtiwcr .gt_left {
#>   text-align: left;
#> }
#> 
#> #lgxvjtiwcr .gt_center {
#>   text-align: center;
#> }
#> 
#> #lgxvjtiwcr .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #lgxvjtiwcr .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #lgxvjtiwcr .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #lgxvjtiwcr .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #lgxvjtiwcr .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #lgxvjtiwcr .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #lgxvjtiwcr .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #lgxvjtiwcr .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #lgxvjtiwcr .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #lgxvjtiwcr .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #lgxvjtiwcr .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #lgxvjtiwcr .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="5" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Loading matrix</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" scope="col" id="Variables">Variables</th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="Factors">
#>         <span class="gt_column_spanner">Factors</span>
#>       </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="2" colspan="1" scope="col" id="Communalities">Communalities</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="2" colspan="1" scope="col" id="Complexity">Complexity</th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="MR1">MR1</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="MR2">MR2</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <th colspan="5" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Loadings">Loadings</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Loadings  Variables" class="gt_row gt_left">qsec</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">-0.92</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">-0.35</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.78</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.29</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">hp</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">0.89</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center"></td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.88</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.04</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">carb</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">0.85</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">0.24</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.66</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.16</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">vs</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">-0.8</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center"></td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.71</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.05</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">cyl</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">0.71</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">-0.48</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.94</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.75</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">mpg</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">-0.61</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">0.54</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.85</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.98</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">am</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center"></td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">0.93</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.82</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.04</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">gear</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">0.22</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">0.93</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.8</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.11</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">drat</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center"></td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">0.78</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.71</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.1</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">wt</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">0.42</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">-0.7</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.82</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">1.64</td></tr>
#>     <tr><td headers="Loadings  Variables" class="gt_row gt_left">disp</td>
#> <td headers="Loadings  MR1" class="gt_row gt_center">0.58</td>
#> <td headers="Loadings  MR2" class="gt_row gt_center">-0.6</td>
#> <td headers="Loadings  Communalities" class="gt_row gt_center">0.9</td>
#> <td headers="Loadings  Complexity" class="gt_row gt_center">2</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="5" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Variances">Variances</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Variances  Variables" class="gt_row gt_left">Eigenvalues</td>
#> <td headers="Variances  MR1" class="gt_row gt_center">4.71</td>
#> <td headers="Variances  MR2" class="gt_row gt_center">4.15</td>
#> <td headers="Variances  Communalities" class="gt_row gt_center"></td>
#> <td headers="Variances  Complexity" class="gt_row gt_center"></td></tr>
#>     <tr><td headers="Variances  Variables" class="gt_row gt_left">Explained variance</td>
#> <td headers="Variances  MR1" class="gt_row gt_center">0.43</td>
#> <td headers="Variances  MR2" class="gt_row gt_center">0.38</td>
#> <td headers="Variances  Communalities" class="gt_row gt_center"></td>
#> <td headers="Variances  Complexity" class="gt_row gt_center"></td></tr>
#>     <tr><td headers="Variances  Variables" class="gt_row gt_left">Cumulative explained variance</td>
#> <td headers="Variances  MR1" class="gt_row gt_center">0.43</td>
#> <td headers="Variances  MR2" class="gt_row gt_center">0.81</td>
#> <td headers="Variances  Communalities" class="gt_row gt_center"></td>
#> <td headers="Variances  Complexity" class="gt_row gt_center"></td></tr>
#>     <tr><td headers="Variances  Variables" class="gt_row gt_left">Proportion explained variance</td>
#> <td headers="Variances  MR1" class="gt_row gt_center">0.53</td>
#> <td headers="Variances  MR2" class="gt_row gt_center">0.47</td>
#> <td headers="Variances  Communalities" class="gt_row gt_center"></td>
#> <td headers="Variances  Complexity" class="gt_row gt_center"></td></tr>
#>     <tr><td headers="Variances  Variables" class="gt_row gt_left">Cumulative proportion explained variance</td>
#> <td headers="Variances  MR1" class="gt_row gt_center">0.53</td>
#> <td headers="Variances  MR2" class="gt_row gt_center">1</td>
#> <td headers="Variances  Communalities" class="gt_row gt_center"></td>
#> <td headers="Variances  Complexity" class="gt_row gt_center"></td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="5"> <em>Note.</em> Extraction method is minres. Rotation method is oblimin. RMSEA is 0.178 CI90% [0.121, 0.245]. Loadings below |0.2| are not displayed.</td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>