Nice html table for lavaan structure equation model fitted objects

Nice html table for lavaan structure equation model fitted objects

nice_sem(
  x,
  standardized = TRUE,
  show_fitmeasures = c(CFI = "cfi", TLI = "tli", RMSEA = "rmsea", `RMSEA ci lower` =
    "rmsea.ci.lower", SRMR = "srmr", AIC = "aic", BIC = "bic"),
  remove_cols = NULL,
  show_ci = TRUE,
  round = 2,
  ...
)

Arguments

x

An object returned from the lavaan sem or cfa function.

show_fitmeasures

A named vector with fit measures.

remove_cols

Either column number or column names to be removed.

show_ci

If TRUE, adds columns with 95% confidence intervals.

...

Further arguments passed to the nice_table() function.

Examples

nice_sem(wmisc::lavaan_fit)
#> <div id="atjcrgnjjo" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#atjcrgnjjo table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #atjcrgnjjo thead, #atjcrgnjjo tbody, #atjcrgnjjo tfoot, #atjcrgnjjo tr, #atjcrgnjjo td, #atjcrgnjjo th {
#>   border-style: none;
#> }
#> 
#> #atjcrgnjjo p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #atjcrgnjjo .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #atjcrgnjjo .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #atjcrgnjjo .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #atjcrgnjjo .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #atjcrgnjjo .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #atjcrgnjjo .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #atjcrgnjjo .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #atjcrgnjjo .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #atjcrgnjjo .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #atjcrgnjjo .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #atjcrgnjjo .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #atjcrgnjjo .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #atjcrgnjjo .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #atjcrgnjjo .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #atjcrgnjjo .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #atjcrgnjjo .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #atjcrgnjjo .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #atjcrgnjjo .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #atjcrgnjjo .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #atjcrgnjjo .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #atjcrgnjjo .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #atjcrgnjjo .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #atjcrgnjjo .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #atjcrgnjjo .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #atjcrgnjjo .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #atjcrgnjjo .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #atjcrgnjjo .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #atjcrgnjjo .gt_left {
#>   text-align: left;
#> }
#> 
#> #atjcrgnjjo .gt_center {
#>   text-align: center;
#> }
#> 
#> #atjcrgnjjo .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #atjcrgnjjo .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #atjcrgnjjo .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #atjcrgnjjo .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #atjcrgnjjo .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #atjcrgnjjo .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #atjcrgnjjo .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #atjcrgnjjo .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #atjcrgnjjo .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #atjcrgnjjo .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #atjcrgnjjo .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #atjcrgnjjo .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="7" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>Structure equation model</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" scope="col" id="parameter">parameter</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="2" colspan="1" scope="col" id="ß">ß</th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="95% CI">
#>         <span class="gt_column_spanner">95% CI</span>
#>       </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="2" colspan="1" scope="col" id="se">se</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="2" colspan="1" scope="col" id="z">z</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="2" colspan="1" scope="col" id="p">p</th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="lower">lower</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="upper">upper</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <th colspan="7" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Latent variables">Latent variables</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Latent variables  parameter" class="gt_row gt_left">remediation → RU_02</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.90</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.88</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.92</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.01</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">91.87</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">remediation → RM_01</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.90</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.88</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.92</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.01</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">91.94</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">remediation → RU_03</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.87</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.85</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.90</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.01</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">73.12</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">remediation → RM_02</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.89</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.87</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.91</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.01</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">83.82</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">development → RE_01</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.91</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.88</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.94</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.02</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">59.10</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">development → RE_02</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.91</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.88</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.94</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.02</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">59.26</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">goal → RZ_02</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.88</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.85</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.90</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.01</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">63.27</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Latent variables  parameter" class="gt_row gt_left">goal → RZ_01</td>
#> <td headers="Latent variables  ß" class="gt_row gt_center">0.91</td>
#> <td headers="Latent variables  lower" class="gt_row gt_center">0.88</td>
#> <td headers="Latent variables  upper" class="gt_row gt_center">0.93</td>
#> <td headers="Latent variables  se" class="gt_row gt_center">0.01</td>
#> <td headers="Latent variables  z" class="gt_row gt_center">72.00</td>
#> <td headers="Latent variables  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="7" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Covariances">Covariances</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Covariances  parameter" class="gt_row gt_left">remediation ↔ development</td>
#> <td headers="Covariances  ß" class="gt_row gt_center">0.59</td>
#> <td headers="Covariances  lower" class="gt_row gt_center">0.52</td>
#> <td headers="Covariances  upper" class="gt_row gt_center">0.65</td>
#> <td headers="Covariances  se" class="gt_row gt_center">0.03</td>
#> <td headers="Covariances  z" class="gt_row gt_center">18.43</td>
#> <td headers="Covariances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Covariances  parameter" class="gt_row gt_left">remediation ↔ goal</td>
#> <td headers="Covariances  ß" class="gt_row gt_center">0.80</td>
#> <td headers="Covariances  lower" class="gt_row gt_center">0.76</td>
#> <td headers="Covariances  upper" class="gt_row gt_center">0.84</td>
#> <td headers="Covariances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Covariances  z" class="gt_row gt_center">40.24</td>
#> <td headers="Covariances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Covariances  parameter" class="gt_row gt_left">development ↔ goal</td>
#> <td headers="Covariances  ß" class="gt_row gt_center">0.74</td>
#> <td headers="Covariances  lower" class="gt_row gt_center">0.70</td>
#> <td headers="Covariances  upper" class="gt_row gt_center">0.79</td>
#> <td headers="Covariances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Covariances  z" class="gt_row gt_center">30.32</td>
#> <td headers="Covariances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="7" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Variances">Variances</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Variances  parameter" class="gt_row gt_left">RU_02</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.19</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.15</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.22</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Variances  z" class="gt_row gt_center">10.58</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RM_01</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.19</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.15</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.22</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Variances  z" class="gt_row gt_center">10.57</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RU_03</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.24</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.20</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.28</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Variances  z" class="gt_row gt_center">11.55</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RM_02</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.21</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.17</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.24</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Variances  z" class="gt_row gt_center">10.94</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RE_01</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.18</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.12</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.23</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.03</td>
#> <td headers="Variances  z" class="gt_row gt_center">6.39</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RE_02</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.18</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.12</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.23</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.03</td>
#> <td headers="Variances  z" class="gt_row gt_center">6.33</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RZ_02</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.23</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.19</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.28</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Variances  z" class="gt_row gt_center">9.59</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">RZ_01</td>
#> <td headers="Variances  ß" class="gt_row gt_center">0.18</td>
#> <td headers="Variances  lower" class="gt_row gt_center">0.13</td>
#> <td headers="Variances  upper" class="gt_row gt_center">0.22</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.02</td>
#> <td headers="Variances  z" class="gt_row gt_center">7.84</td>
#> <td headers="Variances  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">remediation</td>
#> <td headers="Variances  ß" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  lower" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  upper" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.00</td>
#> <td headers="Variances  z" class="gt_row gt_center"><br /></td>
#> <td headers="Variances  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">development</td>
#> <td headers="Variances  ß" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  lower" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  upper" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.00</td>
#> <td headers="Variances  z" class="gt_row gt_center"><br /></td>
#> <td headers="Variances  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Variances  parameter" class="gt_row gt_left">goal</td>
#> <td headers="Variances  ß" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  lower" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  upper" class="gt_row gt_center">1.00</td>
#> <td headers="Variances  se" class="gt_row gt_center">0.00</td>
#> <td headers="Variances  z" class="gt_row gt_center"><br /></td>
#> <td headers="Variances  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="7" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Intercepts">Intercepts</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Intercepts  parameter" class="gt_row gt_left">RU_02</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">2.94</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.75</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">3.13</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.10</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">30.00</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RM_01</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">2.72</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.54</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">2.90</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.09</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">29.54</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RU_03</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">2.69</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.51</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">2.87</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.09</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">29.45</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RM_02</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">2.48</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.32</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">2.65</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.09</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">28.93</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RE_01</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">3.01</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.82</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">3.21</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.10</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">30.13</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RE_02</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">2.83</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.64</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">3.01</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.09</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">29.77</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RZ_02</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">2.72</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.54</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">2.90</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.09</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">29.54</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">RZ_01</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">3.00</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">2.81</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">3.20</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.10</td>
#> <td headers="Intercepts  z" class="gt_row gt_center">30.11</td>
#> <td headers="Intercepts  p" class="gt_row gt_center">&lt;.001</td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">remediation</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  z" class="gt_row gt_center"><br /></td>
#> <td headers="Intercepts  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">development</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  z" class="gt_row gt_center"><br /></td>
#> <td headers="Intercepts  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Intercepts  parameter" class="gt_row gt_left">goal</td>
#> <td headers="Intercepts  ß" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  lower" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  upper" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  se" class="gt_row gt_center">0.00</td>
#> <td headers="Intercepts  z" class="gt_row gt_center"><br /></td>
#> <td headers="Intercepts  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="7" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Modelfit">Modelfit</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Modelfit  parameter" class="gt_row gt_left">N parameters</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">27.00</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">Observations</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">554.00</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">Χ²(17) / p-value</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">142.89</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center">0.00</td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">CFI</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">0.97</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">TLI</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">0.95</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">RMSEA</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">0.12</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">RMSEA ci lower</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">0.10</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">SRMR</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">0.03</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">AIC</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">9644.88</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>     <tr><td headers="Modelfit  parameter" class="gt_row gt_left">BIC</td>
#> <td headers="Modelfit  ß" class="gt_row gt_center">9761.44</td>
#> <td headers="Modelfit  lower" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  upper" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  se" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  z" class="gt_row gt_center"><br /></td>
#> <td headers="Modelfit  p" class="gt_row gt_center"><br /></td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="7"> <em>Note.</em> The estimation method employed in this analysis is Maximum Likelihood (ML). The Non-Linear Minimization, Bounded (nlminb) algorithm was applied for optimization. Missing data were addressed using the Full Information Maximum Likelihood (FIML) approach. The analysis was performed with the lavaan package in R (Yves Rosseel, 2012).</td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>