Create a nicely formatted table

This function takes a data frame and formats it into a nicely formatted HTML table using the gt packages.

nice_table(
  x,
  title = NULL,
  footnote = NULL,
  spanner = header,
  row_group = NULL,
  row_group_order = NULL,
  header = NULL,
  pack = NULL,
  rownames = FALSE,
  file = NULL,
  cols_label = NULL,
  use_labels = TRUE,
  decimals = NULL,
  round = NULL,
  label_na = NULL,
  engine = getOption("wmisc.nice.table.engine"),
  extra = NULL,
  gt = NULL,
  kable = NULL
)

Arguments

x

The data frame to be formatted into a table

title

Title string.

footnote

Add footnote

row_group

List with information on grouping rows row_group = list("Start" = 1:2, "That is the second" = 3:5)

row_group_order

List with information on grouping order.

file

(works only when engine = "gt") If set, an additional file with the table is produced.

cols_label

List with renaming information for columns (old_name = new_name).

decimals

Number of decimals that will be printed.

round

Number of digits to which numbers should be rounded.

extra

Additional arguments passed to kableExtra::kable_classic_2()

gt

Additional arguments passed to gt::gt()

kable

Additional arguments passed to knitr::kable()

Value

A nicely formatted HTML table

Examples

df <- data.frame(
  x = 1:5, y = rnorm(5, mean = 10, sd = 12),
  c = letters[5:1], d = sample(letters, 5)
)
nice_table(
  df,
  title = "A nice title",
  footnote = c("Footnote 1", "Footnote 2"),
  spanner = list("One" = 1:2, "Two" = 3:4),
  row_group = list("Start" = 1:2, "That is the second" = 3:5),
  cols_label = list(x = "First", y = "Second", c = "Third", d = "Fourth"),
  decimals = 1
)
#> <div id="pfgetcifst" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#pfgetcifst table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #pfgetcifst thead, #pfgetcifst tbody, #pfgetcifst tfoot, #pfgetcifst tr, #pfgetcifst td, #pfgetcifst th {
#>   border-style: none;
#> }
#> 
#> #pfgetcifst p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #pfgetcifst .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #pfgetcifst .gt_title {
#>   color: #333333;
#>   font-size: 100%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #pfgetcifst .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #pfgetcifst .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #pfgetcifst .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #pfgetcifst .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #pfgetcifst .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #pfgetcifst .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #pfgetcifst .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #pfgetcifst .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #pfgetcifst .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #pfgetcifst .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: middle;
#> }
#> 
#> #pfgetcifst .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #pfgetcifst .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #pfgetcifst .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 0px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #pfgetcifst .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #pfgetcifst .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #pfgetcifst .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #pfgetcifst .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #pfgetcifst .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #pfgetcifst .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #pfgetcifst .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #pfgetcifst .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #pfgetcifst .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 3px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 3px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #pfgetcifst .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #pfgetcifst .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #pfgetcifst .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #pfgetcifst .gt_left {
#>   text-align: left;
#> }
#> 
#> #pfgetcifst .gt_center {
#>   text-align: center;
#> }
#> 
#> #pfgetcifst .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #pfgetcifst .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #pfgetcifst .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #pfgetcifst .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #pfgetcifst .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #pfgetcifst .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #pfgetcifst .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #pfgetcifst .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #pfgetcifst .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #pfgetcifst .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #pfgetcifst .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #pfgetcifst .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="4" class="gt_heading gt_title gt_font_normal gt_bottom_border" style><strong>Table</strong><br>  <em>A nice title</em></td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="One">
#>         <span class="gt_column_spanner">One</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="Two">
#>         <span class="gt_column_spanner">Two</span>
#>       </th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="First">First</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Second">Second</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Third">Third</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Fourth">Fourth</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <th colspan="4" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="Start">Start</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="Start  x" class="gt_row gt_left">1.0</td>
#> <td headers="Start  y" class="gt_row gt_center">7.6</td>
#> <td headers="Start  c" class="gt_row gt_center">e</td>
#> <td headers="Start  d" class="gt_row gt_center">x</td></tr>
#>     <tr><td headers="Start  x" class="gt_row gt_left">2.0</td>
#> <td headers="Start  y" class="gt_row gt_center">5.3</td>
#> <td headers="Start  c" class="gt_row gt_center">d</td>
#> <td headers="Start  d" class="gt_row gt_center">e</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <th colspan="4" class="gt_group_heading" style="text-align: center;" scope="colgroup" id="That is the second">That is the second</th>
#>     </tr>
#>     <tr class="gt_row_group_first"><td headers="That is the second  x" class="gt_row gt_left">3.0</td>
#> <td headers="That is the second  y" class="gt_row gt_center">8.8</td>
#> <td headers="That is the second  c" class="gt_row gt_center">c</td>
#> <td headers="That is the second  d" class="gt_row gt_center">z</td></tr>
#>     <tr><td headers="That is the second  x" class="gt_row gt_left">4.0</td>
#> <td headers="That is the second  y" class="gt_row gt_center">13.2</td>
#> <td headers="That is the second  c" class="gt_row gt_center">b</td>
#> <td headers="That is the second  d" class="gt_row gt_center">n</td></tr>
#>     <tr><td headers="That is the second  x" class="gt_row gt_left">5.0</td>
#> <td headers="That is the second  y" class="gt_row gt_center">11.1</td>
#> <td headers="That is the second  c" class="gt_row gt_center">a</td>
#> <td headers="That is the second  d" class="gt_row gt_center">w</td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="4"> <em>Note.</em> Footnote 1. Footnote 2.</td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>